refactor(sync-repo-settings): separate bot logic from probot handlers
