+++ packages/sync-repo-settings/package.json
@@ -33,6 +33,7 @@
   },
   "devDependencies": {
     "@octokit/openapi-types": "^5.1.0",
+    "@octokit/rest": "^18.3.5",
     "@types/extend": "^3.0.1",
     "@types/js-yaml": "^4.0.0",
     "@types/mocha": "^8.0.0",
